% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abalone.R
\docType{data}
\name{abalone}
\alias{abalone}
\title{Imbalanced binary abalone dataset}
\format{
A data frame with 731 rows and 8 variables:
\describe{
  \item{Length}{Real [0.075, 0.815]}
  \item{Diameter}{Real [0.055, 0.65]}
  \item{Height}{Real [0.0, 1.13]}
  \item{Whole_weight}{Real [0.0020, 2.8255]}
  \item{Shucked_weight}{Real [0.0010, 1.488]}
  \item{Viscera_weight}{Real [5.0E-4, 0.76]}
  \item{Shell_weight}{Real [0.0015, 1.005]}
  \item{Class}{Nominal \{positive, negative\}}
  ...
}
}
\source{
\url{KEEL}{https://sci2s.ugr.es/keel/datasets.php}
}
\usage{
abalone
}
\description{
A imbalanced version of the Abalone data set, where the possitive examples
belong to class 18 and the negative examples belong to class 9.
}
\keyword{datasets}
