% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUS.R
\name{RUS}
\alias{RUS}
\title{Random Under-Sampling Algorithm}
\usage{
RUS(data, outcome, perc_min = 100)
}
\arguments{
\item{data}{A dataset containing the predictors and the outcome. The predictors
can be continuous (\code{numeric} or \code{integer}) or catigorical (\code{character}
or \code{factor}). The outcome must be binary.}

\item{outcome}{The column number or the name of the outcome variable in the dataset.}

\item{perc_min}{The desired percentage of the size of minority samples that the
majority samples would be reached in the new dataset. The default is 100.}
}
\value{
A new dataset has been balanced.
}
\description{
A balanced dataset would be return by using random under-sampling (RUS) algorithm.
}
\details{
The random under-sampling algorithm randomly chooses the majority samples without
replacement according to the sample size of minority class, in order to get a
more balanced dataset.
}
\examples{
data(abalone)
table(abalone$Class)

newdata1 <- RUS(abalone, 'Class')
table(newdata1$Class)

newdata2 <- RUS(abalone, 'Class', perc_min=200)
table(newdata2$Class)
}
