% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBC.R
\name{SBC}
\alias{SBC}
\title{Under-Sampling Based on Clustering Algorithm}
\usage{
SBC(data, outcome, perc_min = 100, k = 3, iter_max = 100,
  nstart = 1, ...)
}
\arguments{
\item{data}{A dataset containing the predictors and the outcome. The predictors
can be continuous (\code{numeric} or \code{integer}) or catigorical (\code{character}
or \code{factor}). The outcome must be binary.}

\item{outcome}{The column number or the name of the outcome variable in the dataset.}

\item{perc_min}{The desired percentage of the size of minority samples that the
majority samples would be reached in the new dataset. The default is 100.}

\item{k}{The number of clusters for the clustering algorithm. The default is 3.}

\item{iter_max}{The maximum number of iterations of the clustering algorithm.
The default is 100.}

\item{nstart}{The initial number of random sets would be chosen. Only would be
used for \code{k-means} and \code{k-prototypes}. The default is 1.}

\item{...}{Not used.}
}
\value{
A new dataset has been balanced.
}
\description{
A balanced dataset would be return by using under-sampling based on clustering
(SBC) algorithm.
}
\details{
The under-sampling based on clustering algorithm clusters all samples into
\code{k} clusters. Then it randomly selects the majority samples by considering
the ratio of the number of majority samples to the number of minority samples
in the cluster.

If we need to sample more majority samples than what is available in the cluster,
the sampling with replacement would be used. Otherwise, the sampling without
replacement would be used.

For the dataset with predictors that are all continuous (\code{numeric} or
\code{integer}), \code{k-means} would be used to cluster. For the dataset with
predictors that are all categorical (\code{character} or \code{factor}),
\code{k-modes} would be used. For the dataset with predictors are continuous or
categorical, \code{k-prototypes} would be used.
}
\examples{
data(abalone)
table(abalone$Class)

# predictors are continuous or categorical
newdata1 <- SBC(abalone, 'Class')
table(newdata1$Class)

newdata2 <- SBC(abalone, 'Class', perc_min=200)
table(newdata2$Class)

# predictors are all continuous
newdata3 <- SBC(abalone[, -1], 'Class')
table(newdata3$Class)

# predictors are all categorical
abalone1 <- abalone
abalone1[, 2:8] <- apply(abalone1[, 2:8], 2, function(x) ifelse(x > 0.5, 'high', 'low'))
newdata4 <- SBC(abalone1, 'Class')
table(newdata4$Class)
}
\references{
Yen, S. J., & Lee, Y. S. (2009). Cluster-based under-sampling
approaches for imbalanced data distributions. \emph{Expert Systems with Applications},
36(3), 5718-5727.

Hartigan, J. A., & Wong, M. A. (1979). Algorithm AS 136: A k-means
clustering algorithm. \emph{Journal of the Royal Statistical Society. Series C
(Applied Statistics)}, 28(1), 100-108.

Huang, Z. (1997). A fast clustering algorithm to cluster very large
categorical data sets in data mining. \emph{DMKD}, 3(8), 34-39.

Huang, Z. (1998). Extensions to the k-means algorithm for clustering
large data sets with categorical values. \emph{Data Mining and Knowledge Discovery},
2(3), 283-304.
}
