% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMOTE.R
\name{SMOTE}
\alias{SMOTE}
\title{Synthetic Minority Over-sampling TEchnique}
\usage{
SMOTE(data, outcome, perc_maj = 100, k = 5)
}
\arguments{
\item{data}{A dataset containing the predictors and the outcome. The predictors
can only be continuous (\code{numeric} or \code{integer}). The outcome must be binary.}

\item{outcome}{The column number or the name of the outcome variable in the dataset.}

\item{perc_maj}{The desired percentage of the size of majority samples that the
minority samples would be reached in the new dataset. The default is 100.}

\item{k}{The number of nearest neighbours that are used to generate the new samples
of the minority class. The default is 5.}
}
\value{
A new dataset has been balanced.
}
\description{
A balanced dataset would be return by using Synthetic Minority Over-sampling
TEchnique (SMOTE) algorithm.
}
\details{
The synthetic minority over-sampling technique artificially generates new samples
of the minority class using the nearest neighbours of these cases, in order to get a
more balanced dataset.
}
\examples{
data(abalone)
table(abalone$Class)

newdata1 <- SMOTE(abalone, 'Class')
table(newdata1$Class)

newdata2 <- SMOTE(abalone, 'Class', perc_maj=50)
table(newdata2$Class)
}
\references{
Chawla, N. V., Bowyer, K. W., Hall, L. O., & Kegelmeyer, W. P. (2002).
SMOTE: synthetic minority over-sampling technique. \emph{Journal of artificial
intelligence research}, 16, 321-357.
}
