% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMOTE_NC.R
\name{SMOTE_NC}
\alias{SMOTE_NC}
\title{Synthetic Minority Over-sampling TEchnique-Nominal Continuous}
\usage{
SMOTE_NC(data, outcome, perc_maj = 100, k = 5)
}
\arguments{
\item{data}{A dataset containing the predictors and the outcome. The predictors
can be continuous (\code{numeric} or \code{integer}) or catigorical (\code{character}
or \code{factor}). There must be at least one continuous predictor and at least
one categorical predictor. The outcome must be binary.}

\item{outcome}{The column number or the name of the outcome variable in the dataset.}

\item{perc_maj}{The desired percentage of the size of majority samples that the
minority samples would be reached in the new dataset. The default is 100.}

\item{k}{The number of nearest neighbours that are used to generate the new samples
of the minority class. The default is 5.}
}
\value{
A new dataset has been balanced.
}
\description{
A balanced dataset would be return by using Synthetic Minority Over-sampling
TEchnique-Nominal Continuous (SMOTE-NC) algorithm.
}
\details{
The synthetic minority over-sampling technique-nominal continuous artificially
generates new samples of the minority class using the nearest neighbours of
these cases, in order to get a more balanced dataset. This algorithm could
handle mixed datasets of continuous and nominal features, but it could not
handle datasets with all nominal features or all continuous features.
}
\examples{
data(abalone)
table(abalone$Class)

newdata1 <- SMOTE_NC(abalone, 'Class')
table(newdata1$Class)

newdata2 <- SMOTE_NC(abalone, 'Class', perc_maj=50)
table(newdata2$Class)
}
\references{
Chawla, N. V., Bowyer, K. W., Hall, L. O., & Kegelmeyer, W. P. (2002).
SMOTE: synthetic minority over-sampling technique. \emph{Journal of artificial
intelligence research}, 16, 321-357.
}
